refactor: move units fns into own directory
